+++ Scripts/MainMenu.gd
@@ -6,11 +6,21 @@ signal credits_pressed
 signal cheats_pressed
 signal game_mode_pressed
 signal statistics_pressed
+signal options_pressed
 
 # Called when the node enters the scene tree for the first time.
 func _ready():
     $CanvasLayer/MainMenuContainer/StartGame.grab_focus()
     $CanvasLayer/VersionLabel.text = "Version " + constants.GAME_VERSION
+    
+    if OS.has_feature('steam'):
+        $CanvasLayer/VersionLabel.text += " (Steam)"
+    
+    # Disable Statistics and Exit Game for web version.
+    if OS.has_feature('web'):
+        $CanvasLayer/MainMenuContainer/Statistics.visible = false
+        $CanvasLayer/MainMenuContainer/Options.visible = false
+        $CanvasLayer/MainMenuContainer/ExitGame.visible = false
     
 # Called every frame. 'delta' is the elapsed time since the previous frame.
 func _process(_delta):
@@ -35,3 +45,6 @@ func _on_game_mode_pressed():
 
 func _on_statistics_pressed():
     emit_signal('statistics_pressed')
+
+func _on_options_pressed():
+    emit_signal('options_pressed')