+++ Scripts/Storage.gd
@@ -2,14 +2,38 @@ extends Node
 
 
 var Config
-
 var Stats
+
+# -- SYSTEM CONFIGURATION --
+# Per-system.  Should NOT be Cloud Saved.  And so deliberately a different file.
+
+func load_config():
+    Config = ConfigFile.new()
+    
+    var err = Config.load("user://config.ini")
+
+    if err != OK:
+        Config.set_value('config','screen_mode','FULL_SCREEN')
+        Config.set_value('config','master_volume',1.0)
+        Config.set_value('config','music_volume',1.0)
+        Config.set_value('config','effects_volume',1.0)
+
+func save_config():
+    if OS.has_feature('web'):
+        return
+
+    var err = Config.save("user://config.ini")
+    
+    if err != OK:
+        print("config(): Fail")
+
+# -- STATISTICS --
 
 func load_stats():
     Stats = ConfigFile.new()
     
-    var err = Stats.load("user://stats.ini")
-        
+    var err = Stats.load("user://cloud-stats.ini")
+       
     if err != OK:
         # Could not load stats.  That's OK, might be first run.
         Stats.set_value('player','high_score',0)
@@ -19,10 +43,11 @@ func load_stats():
         Stats.set_value('player','furthest_wave', 0)
         Stats.set_value('player','fish_rescued', 0)
         
+func save_stats():
+    if OS.has_feature('web'):
         return
         
-func save_stats():
-    var err = Stats.save("user://stats.ini")
+    var err = Stats.save("user://cloud-stats.ini")
     
     if err != OK:
         print("save_stats(): Fail")