+++ CHANGELOG.md
@@ -1,6 +1,52 @@
 # Changelog
 
-# 0.3-alpha
+# 0.4-alpha (2023-05-12)
+
+In this release, the gameplay has been sped up to make it more frenetic, combined
+with the addition of skeletons and GRENADES!  The HUD has been streamlined, and the player has a 
+targeting laser when playing with a controller. 
+
+Away from the core game, we now also have an Options screen and Steam Cloud
+support.
+
+* Gameplay:
+    * Increase player fire rate and re-balance FAST SPRAY
+    * Enemies can no longer take damage while they are spawning.
+    * Increase enemy spawn numbers.
+    * % chance of multiple spawns happening at the same time (Will be different types)
+    * Increase knight health to 4 (was: 1)
+    * Increase necromancer health to 10 (was: 3)
+    * Increase trap health to 10 (was: 5)
+* Enemies:
+    * Added Skeletons (that split into 4 when vanquished!)
+* Power-ups:
+    * Added GRENADE power-up (Boom!)
+    * Power-ups no longer 'tick' unless the player is actively controlling the shark.
+* Controls:
+    * Controller: Player now has a targetting laser! 
+    * Controller: 'B' button to back out of menus.
+* Graphics:
+    * Power up bar moved to RHS.  Non-activated power-ups are not shown.
+* Sound:
+    * Master / Music / Volume audio buses implemented.
+* Misc
+    * Options screen added.  Persistent storage used.
+        * Screen mode (Full screen, windowed)
+        * Volume (Master, Music, Effects)
+    * Credits screen re-jigged.
+    * Pacifist Mode: Hide power-up bar.
+    * Windowed Mode: Use 1920x1080 as a default (User can re-size)
+* Platform:
+    * HTML5: Disable storage, Statistics and Exit Game.
+    * Steam: Steam Cloud sync for Statistics (cross-platform)
+* Bug fixes
+    * Player projectiles would not pass through spawning enemies.
+    * Magnet: Make a bit stronger to avoid items chasing the player(!)
+    * Controller START button mapped to too many menu items (and so not working).
+    * Steam Deck: STATISTICS not consistently working when selected with controller.
+    * Steam Deck: CREDITS not consistently working when selected with controller.
+
+# 0.3-alpha (2023-05-03)
 
 This release introduces the upgrade system, a new Pacifist game mode, more interesting wave /
 spawning mechanics and much more!
@@ -61,7 +107,7 @@ spawning mechanics and much more!
     * Improve arena spawn co-ordinates range to prevent insta-collisions.
     * Fixed shonky calculations for enemy spawn numbers within a wave.
 
-# 0.2-alpha
+# 0.2-alpha (2023-04-16)
 
 This is primarily a bug fix / 'quality of life' release.
 
@@ -76,6 +122,6 @@ This is primarily a bug fix / 'quality of life' release.
 * Bug fix: Make 'Enemies Left' behaviour consistent between games.
 * Bug fix: Tidy up projectiles getting stuck around exit locations in transitions.
 
-# 0.1-alpha
+# 0.1-alpha (2023-04-06)
 
 First release.